Fix locale related tests
Some tests are skipped when the required locale is not available.  This
patch installs all of the required locales for CI on Travis and Azure.

There were several broken tests.  Two gettext tests needed
`putenv("LANGUAGE={locale}");`.  The reason is pointed out by one of the
comments in the manual page for gettext:

> If [the environment] has a non-empty LANGUAGE, you may end up with
> unexpected localization strings.
>
> [Because] when a program looks up locale dependent values, it does this
> according to the following environment variables, in priority order:
>
>     LANGUAGE
>     LC_ALL
>     LC_xxx according to selected locale category: LC_CTYPE, LC_NUMERIC,
>            LC_TIME, LC_COLLATE, LC_MONETARY, LC_MESSAGES, ...
>     LANG
>
> https://www.php.net/manual/en/function.gettext.php

A postgresql test that removed because it was failing due the new
"Locale-independent float to string cast" changes in 8.0.
See https://wiki.php.net/rfc/locale_independent_float_to_string

The reason the test wasn't failing in CI environments is because it was
skipped due to a missing locale requirement.

I also updated some of the skip reason messages because it was difficult
to see why the tests were being skipped.